print_results: Remove unneeded exception throwing
This method is for coala internal use only and manually type checking
every parameter is a thing of the past. We want our code to fail if
something doesn't work out and python does this on its own, this is an
unneeded maintenance overhead in terms of code and testing.